refactored IO tests
* separated *IOTest in *DataSourceTest and *DataSinkTest
* renamed HBaseIOTest to HBaseDataSourceSinkTest to re-use HBase/HDFS mini cluster
* fixes #339
