+++ gradoop-flink/src/test/resources/data/dot/input.gdl
@@ -1,4 +1,4 @@
-singleGraph:Network{from="2006"}[
+input:Network{from="2006"}[
 (a:Person{name="Anna",age="20"})-[:knows{since="2012"}]->(b:Person{name="Tom",age="25"})
 (b)-[:knows{since="2012"}]->(a)
 (c:Forum{topic="Apache Flink",lan="EN"})-[:hasModerator{since="2002"}]->(b)
